fix: columnView, columnVisibility 초기화 위치 변경
