+++ client/src/hooks/store/useColumnView.tsx
@@ -1,26 +1,10 @@
-import { MouseEventHandler, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
+import { MouseEventHandler } from 'react';
 import { useRecoilState } from 'recoil';
 
 import { columnViewAtom } from '@/store/atom/columnView';
-import { dataframeAtom } from '@/store/atom/dataframe';
 
-const useColumnView = (idx?: number, update = true) => {
-  const [{ columns }] = useRecoilState(dataframeAtom);
-
+const useColumnView = (idx?: number) => {
   const [columnView, setColumnView] = useRecoilState(columnViewAtom);
-
-  const { id } = useParams();
-
-  useEffect(() => {
-    if (!update) return;
-
-    const initialState = Array.from({ length: columns.length }).fill(
-      true,
-    ) as boolean[];
-
-    setColumnView(initialState);
-  }, [id]);
 
   if (idx === undefined) return { columnView };
